requirements.txt
structpy==0.2
lark-parser
regex
pytz
